refactor(rename): 도메인별 디렉토리 생성 및 이동
