+++ src/main/java/qna/domain/AbstractEntity.java
@@ -3,6 +3,8 @@ package qna.domain;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+import qna.domain.deleteHistory.DeleteHistory;
+import qna.domain.user.User;
 
 import javax.persistence.Column;
 import javax.persistence.EntityListeners;
