+++ src/main/java/qna/domain/answer/Answer.java
@@ -1,7 +1,12 @@
-package qna.domain;
+package qna.domain.answer;
 
 import qna.NotFoundException;
 import qna.UnAuthorizedException;
+import qna.domain.AbstractEntity;
+import qna.domain.ContentType;
+import qna.domain.deleteHistory.DeleteHistory;
+import qna.domain.question.Question;
+import qna.domain.user.User;
 
 import javax.persistence.Entity;
 import javax.persistence.ForeignKey;
