+++ src/main/java/qna/domain/deleteHistory/DeleteHistory.java
@@ -1,6 +1,16 @@
-package qna.domain;
+package qna.domain.deleteHistory;
 
-import javax.persistence.*;
+import qna.domain.ContentType;
+import qna.domain.user.User;
+
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.ForeignKey;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
 import java.time.LocalDateTime;
 import java.util.Objects;
