+++ src/test/java/qna/domain/answer/AnswersTest.java
@@ -1,7 +1,12 @@
-package qna.domain;
+package qna.domain.answer;
 
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import qna.domain.ContentType;
+import qna.domain.deleteHistory.DeleteHistory;
+import qna.domain.question.Question;
+import qna.domain.user.User;
+import qna.domain.user.UserTest;
 
 import java.time.LocalDateTime;
 import java.util.List;
