+++ src/test/java/qna/service/QnaServiceTest.java
@@ -7,14 +7,14 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import qna.CannotDeleteException;
-import qna.domain.Answer;
 import qna.domain.ContentType;
-import qna.domain.DeleteHistories;
-import qna.domain.DeleteHistory;
-import qna.domain.Question;
-import qna.domain.QuestionRepository;
-import qna.domain.QuestionTest;
-import qna.domain.UserTest;
+import qna.domain.answer.Answer;
+import qna.domain.deleteHistory.DeleteHistories;
+import qna.domain.deleteHistory.DeleteHistory;
+import qna.domain.question.Question;
+import qna.domain.question.QuestionRepository;
+import qna.domain.question.QuestionTest;
+import qna.domain.user.UserTest;
 
 import java.time.LocalDateTime;
 import java.util.Arrays;
